# docs/openapi/famlin.yaml
...
openapi: 3.0.3
info:
  title: Famlin API
  version: current
  description: |
    Every Famlin server exposes this REST API under `/api` — the same API the
    official mobile and web apps use. With a [personal access token](/developers/api)
    you can build your own integrations against your family's server.

    All content is scoped by **group membership**: you can only read and write
    content in groups (families) you belong to, and there is no way around that
    per token. See [Using the API](/developers/api) for authentication,
    conventions, and worked examples.
servers:
  - url: https://famlin.example.com
    description: Your own Famlin deployment
security:
  - bearerAuth: []
...
paths:
...
  /api/posts:
    get:
      operationId: listPosts
      tags: [Posts]
      summary: List posts (your feed)
      description: |
        Newest first, cursor-paginated. The feed is a filter over your groups:
        `groupIds` selects a subset, the legacy `groupId` selects one, and
        neither means every group you belong to. Requesting a group you're not
        a member of is a `403`.
      parameters:
        - name: groupIds
          in: query
          schema: { type: string }
          description: Comma-separated group ids (must all be groups you're in).
          example: cmb1ghi,cmb1jkl
        - name: groupId
          in: query
          schema: { type: string }
          description: Legacy single-group filter.
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/take'
      responses:
        '200':
          description: A page of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
    post:
      operationId: createPost
      tags: [Posts]
      summary: Create a post
      description: |
        `uploadedAssetUrls` must be paths returned by the
        [upload endpoint](/developers/api-reference/upload-files) (`/uploads/...`)
        or Immich proxy paths for an album linked to **this** post's group —
        arbitrary URLs are rejected. `latitude` and `longitude` must be
        provided together.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [groupId]
              properties:
                groupId: { type: string }
                content:
                  type: string
                  maxLength: 5000
                type:
                  type: string
                  enum: [UPDATE, MILESTONE]
                  default: UPDATE
                milestoneTag:
                  type: string
                  maxLength: 50
                  example: First steps!
                uploadedAssetUrls:
                  type: array
                  maxItems: 20
                  items: { type: string }
                  example: ['/uploads/8d0f3c9a-1b2e-4f5a-9c8d-7e6f5a4b3c2d.jpg']
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
                locationName:
                  type: string
                  maxLength: 200
      responses:
        '200':
          description: The created post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
  /api/posts/search:
    get:
      operationId: searchPosts
      tags: [Posts]
      summary: Search posts in a group
      description: Case-insensitive match on `content` and `milestoneTag`, within one group.
      parameters:
        - name: groupId
          in: query
          required: true
          schema: { type: string }
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - $ref: '#/components/parameters/cursor'
        - $ref: '#/components/parameters/take'
      responses:
        '200':
          description: A page of matching posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostsPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
  /api/posts/on-this-day:
    get:
      operationId: getOnThisDay
      tags: [Posts]
      summary: Posts from this day in earlier years
      description: |
        Posts created on today's month/day in a prior year. Not paginated —
        the result set is at most a handful of posts.
      parameters:
        - name: groupId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Memories
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
  /api/posts/{id}:
    get:
      operationId: getPost
      tags: [Posts]
      summary: Get one post
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: The post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/NotGroupMember'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updatePost
      tags: [Posts]
      summary: Edit your own post
      description: |
        Only the author can edit, and only while still a member of the post's
        group. Sets `editedAt`. Pass `null` for the location fields to clear
        the location.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  maxLength: 5000
                milestoneTag:
                  type: string
                  maxLength: 50
                latitude:
                  type: number
                  nullable: true
                longitude:
                  type: number
                  nullable: true
                locationName:
                  type: string
                  nullable: true
      responses:
        '200':
          description: The updated post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not the author, or no longer a member of the group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deletePost
      tags: [Posts]
      summary: Delete a post
      description: |
        Permanent — cascades to the post's comments, reactions, and favorites.
        **There is no undo.** Only the author (or an admin) can delete.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not the author or an admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A personal access token (`famlin_pat_...`, created in the web app) or
        a session JWT from `POST /api/auth/login`. Both are sent the same way:
        `Authorization: Bearer <token>`.
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Cursor from the previous page's `nextCursor`.
    take:
      name: take
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 30
      description: Page size.
  responses:
    Unauthorized:
      description: Missing, invalid, expired, or revoked credential
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotGroupMember:
      description: You are not a member of this group
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Invalid request body or query (see `details`)
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: object
                    description: Zod field errors.
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable, translated via the `Accept-Language` header (`en`/`nl`).
      required: [error]
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ReactionType:
      type: string
      enum: [LIKE, LOVE, HAHA, WOW, SAD, CARE]
      default: LIKE
    ReactionCounts:
      type: object
      description: Per-emoji reaction counts; only emojis with at least one reaction appear.
      additionalProperties:
        type: integer
      example:
        LOVE: 3
        LIKE: 2
...
    UserSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        avatarUrl:
          type: string
          nullable: true
...
    Post:
      type: object
      properties:
        id: { type: string }
        authorId: { type: string }
        author:
          $ref: '#/components/schemas/UserSummary'
        groupId: { type: string }
        group:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
        content:
          type: string
          nullable: true
          example: First day of school!
        type:
          type: string
          enum: [UPDATE, MILESTONE]
        milestoneTag:
          type: string
          nullable: true
        uploadedAssetUrls:
          type: array
          items: { type: string }
          description: '`/uploads/...` or `/api/immich/assets/...` paths — auth-gated, see Uploads & media.'
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        locationName:
          type: string
          nullable: true
        createdAt: { type: string, format: date-time }
        editedAt:
          type: string
          format: date-time
          nullable: true
        commentCount: { type: integer }
        likeCount: { type: integer }
        likedByMe: { type: boolean }
        myReaction:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ReactionType'
        reactions:
          $ref: '#/components/schemas/ReactionCounts'
        recentReactors:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/UserSummary'
          description: The three most recent reactors, newest first.
        favoritedByMe: { type: boolean }
    PostsPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        nextCursor:
          type: string
          nullable: true
          description: Pass back as `?cursor=` for the next page; `null` on the last page.
...